requirements.txt
attrs==23.1.0
Automat==22.10.0
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
constantly==15.1.0
cryptography==41.0.4
cssselect==1.2.0
filelock==3.12.4
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.3
messagebox==0.1.0
mutagen==1.47.0
packaging==23.2
parsel==1.8.1
Pillow==10.1.0
Protego==0.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.19.0
PyDispatcher==2.0.7
pygame==2.5.2
pyOpenSSL==23.2.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.11.0
service-identity==23.1.0
six==1.16.0
tk==0.1.0
tldextract==5.0.1
Twisted==22.10.0
typing_extensions==4.8.0
urllib3==2.0.7
w3lib==2.1.2
websockets==12.0
yt-dlp==2023.10.13
zope.interface==6.1
